Address Bugbot research template findings
