+++ typescript/deep-research-agent/package.json
@@ -9,7 +9,7 @@
     "bench": "tsx src/research.ts",
     "dev": "tsx watch src/research.ts",
     "dev:web": "tsx watch src/local-web.ts",
-    "check": "tsc --noEmit --target ES2022 --module NodeNext --moduleResolution NodeNext --skipLibCheck index.ts src/research.ts src/local-web.ts api/health.ts api/research.ts"
+    "check": "tsc --noEmit --target ES2022 --module NodeNext --moduleResolution NodeNext --skipLibCheck index.ts src/research.ts src/api-utils.ts src/local-web.ts api/health.ts api/research.ts"
   },
   "dependencies": {
     "@browserbasehq/sdk": "^2.9.0",
